mkJobset will set the -src input for all sources sharing the same url.
The gitTree discovered revisions are identified by the URL of the
repository.  The -src inputs are for each potential source.  A single
repository may host several packages (via the subpath specification)
so this change ensures that all of those package -src references to
the same repository are for the same revision.