Allow missing subpath sources (for new modules in a submodule repo).
Adding a new repo in a repository that supports multiple
repos (modules) will break most existing builds because the
configs.nix input is updated to describe and require that input, but
the submodules-specified version of the repository doesn't have that
input yet.  Since the submodule version will not use the new repo,
create a dummy placeholder that will allow evaluation but not build.